Extract methods from ViewController
