+++ Launcher/Classes/DeviceCollector.swift
@@ -45,4 +45,16 @@ class DeviceCollector {
         }
         self.outputPipe.fileHandleForReading.waitForDataInBackgroundAndNotify()
     }
+    
+    // FIXME: Use regular expression here
+    func getDeviceUDID(device : String) -> String {
+        let simulatorUDIDs = device
+            .split(separators: "[]")
+            .map(String.init)
+        if simulatorUDIDs.count >= 2 {
+            return simulatorUDIDs[1]
+        }
+        return ""
+    }
+    
 }
